requirements.txt
# Synthetic Khmer ID card generator – Python dependencies
# Install with:  pip install -r requirements.txt

Pillow>=9.0.0           # image rendering and textbbox() support
albumentations>=1.3.0   # photometric augmentation pipeline
opencv-python>=4.6.0    # used by albumentations internally
numpy>=1.23.0

# --- upload_to_huggingface.py dependencies ---
datasets>=2.14.0        # HF Dataset builder + push_to_hub
huggingface_hub>=0.20.0 # HfApi, CommitOperationAdd
tqdm>=4.65.0            # progress bars

# --- train_yolo_mlflow.py dependencies ---
ultralytics>=8.3.0      # YOLO11 / YOLOv8 training engine
mlflow>=2.13.0          # experiment tracking server + client
tabulate>=0.9.0         # pretty-print leaderboard table
PyYAML>=6.0             # read Ultralytics args.yaml

# --- inference.py dependencies ---
streamlit>=1.35.0       # Streamlit web UI